+++ CHANGELOG.md
@@ -181,5 +181,6 @@
 
 ## 3.1.2
 * Added one more day to Dark Sky forecast
+* Improved error logging for Weather Underground
 * Fixed undefined this.log when converting unkown weather categories
 * Fixed RainBool threshold too strict in DarkSky service